Extract Nav stateless component
